refactor(backend-kit): promote MetricCard to shared kit
Network console becomes the kit's documented reference look. The network
copy is now a re-export shim, so analytics/devices/health render identically
(render body verified byte-identical). Adds a note that semantic colour
(aging buckets, channels, day-done, severity) survives restyles.
